Refactored a few minor thigns to make it easier to port logic to the free synced patterns for themes plugin
