+++ src/PatternBuilder_EditorTools.js
@@ -2,14 +2,13 @@
  * WordPress dependencies
  */
 import { registerPlugin } from '@wordpress/plugins';
-import { addFilter } from '@wordpress/hooks';
 
 /**
  * Internal dependencies
  */
 import { EditorSidePanel } from './components/EditorSidePanel';
 import { PatternPanelAdditionsPlugin } from './components/PatternPanelAdditions';
-import { syncedPatternFilter } from './utils/syncedPatternFilter';
+import './utils/syncedPatternFilter';
 
 registerPlugin('pattern-builder-editor-side-panel', {
 	render: EditorSidePanel,
@@ -18,9 +17,3 @@ registerPlugin('pattern-builder-editor-side-panel', {
 registerPlugin( 'pattern-builder-pattern-panel-additions', {
 	render: PatternPanelAdditionsPlugin,
 } );
-
-addFilter(
-	'editor.BlockEdit',
-	'pattern-builder/pattern-edit',
-	syncedPatternFilter
-);
